+++ app/src/main/res/layout/list_item.xml
@@ -1,5 +1,5 @@
 <?xml version="1.0" encoding="utf-8"?>
-<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
+<com.example.khalk.network.CustomLinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
     android:id="@+id/item_container"
     android:layout_width="match_parent"
     android:layout_height="wrap_content"
@@ -11,6 +11,12 @@
         android:layout_height="wrap_content"
         android:paddingRight="8dp"
         android:text="@string/audio_08" />
+    <TextView
+        android:id="@+id/expected_code"
+        android:layout_width="50dp"
+        android:layout_height="wrap_content"
+        android:paddingRight="8dp"
+        android:text="expected" />
 
     <Button
         android:id="@+id/test_button"
@@ -33,4 +39,4 @@
         android:layout_height="wrap_content"
         android:paddingLeft="16dp" />
 
-</LinearLayout>
+</com.example.khalk.network.CustomLinearLayout>
